Rename scaffold to runtime
scaffold implies that we only generate code, but we also need
to patch files (link them to tasks), execute files (using local tools)
and lookup the airplane special comment in files.
